Update 5e-shaped-scripts.js
this is now working on roll 20 after the changes to the API the only problem i ran into is the cost on Legendary. i had to do a work around on spells it will no longer look in traits, this is working for now but i don't know it it will hold up for all monsters in the database and Lair Actions are not going in at all for now but i don't think that was working in the original script i had. i understand this code is dirty but it works for now